fix(dashboard): rebuild the XvB decision table as per-tier blocks (#1316)
The table was six columns inside a `.table-scroll` in a `.grid-columns` card, and
never showed a whole row. Measured by rendering the real component in the
OPERATOR'S state (XvB off, so build_xvb_calc serves the vendored published
fallback, realization is None and the odds feed has nothing) — not with XvB on,
which is how #991 got its reorder wrong. The widest cell was a Net band at 33
characters and the whole row came to 115.

Four fixes, smallest first:

1. coinDp compared the signed value, so every negative failed both `>=` tests
   and fell through to 8 dp. A loss was always the longest string in its row and
   carried four trailing zeros that said nothing. Now compares magnitude.
   coinTriplet is unaffected — it only ever passes positives.

2. Bands collapse to their midpoint; the band moves into the cell's tooltip, so
   the spread is one hover away rather than gone. The fiat line collapses too.

3. The face-value net is computed in its own right (`netFace`), not only as a
   fallback, and shown beside the study/measured net. The reader had to subtract
   two adjacent columns by eye to get it.

4. A net band straddling zero now renders `status-warn`. As plain uncoloured text
   it was indistinguishable from a cell carrying no verdict at all. The
   `c-purple` accent comes off the XvB-says figure, which had been colouring
   every row regardless of sign — three unrelated colour meanings in one row.

Why the layout changed too, rather than stopping at 1-3 as the issue allowed.
Collapsing the ranges takes the row from 115 characters to 83, which still needs
530-830px depending on the glyph advance (0.95rem tabular digits are ~8-9.1px;
530px assumes an implausibly narrow face). The card's content box is 308px in a
minimum-width column, 411px at the app's own 1400px max container (3 columns of
453px), and at most 492px anywhere in the two-column band. The shortened table
fits none of them. Only the degenerate single-column window (677px) is close,
and only with a font nothing here uses.

Rejected alternative, for the record: `table-layout: fixed` plus dropping
`white-space: nowrap`. Six fixed columns across 308-411px is 51-68px each,
about six characters — every cell becomes a stack of wrapped fragments and the
tier name alone takes three lines. It stops the panning by making the table
unreadable.

So each tier gets a block: the two net verdicts in the existing two-up
`.stat-grid`, their inputs on a plain `<p>` beneath. That paragraph wraps
natively, which is the whole point — the failure being fixed was a line that
would not wrap. It needs NO new CSS: the tier name reuses `.est-heading` and the
separators are real `·` characters, which survive copy-paste and a screen reader
where a flex `gap` would not.

The XvB code moves into its own `xvblogic.mjs` / `xvbview.mjs` pair with its own
`xvbview.test.mjs`, matching the workerlogic/workerview convention. That is not
tidying: the file-budget ratchet only ever lowers ceilings, so growing
components.mjs and logic.mjs was not an option. Both files and both test files
now sit further below their ceilings than before. `EstTable` comes out into
`esttable.mjs` rather than being copied into the new module — it is "the one
shape every earnings tab presents its estimate in", and two copies would have
drifted into two earnings tables that disagree.

Coverage: the operator's XvB-off state gets an explicit regression test that no
visible figure is a range and no 8-dp negative renders; the warn verdict, the
face net, coinDp's magnitude rule and the collapsed fiat line each get their own.

Not verified: there is no browser on this host, so the widths above are derived
from the CSS and the rendered strings, not measured from a live layout.
