+++ dashboard/mining_dashboard/web/static/components.mjs
@@ -6,14 +6,12 @@
 import { BackupPanel } from "./backupview.mjs";
 import { ChartCard } from "./chart.mjs";
 import { ConfigView, UpgradeControl } from "./configview.mjs";
+import { EstTable } from "./esttable.mjs";
 import {
   coinFiat,
-  coinTriplet,
   computeEarnings,
   computeEnergy,
-  computeXvbTier,
   egressRoute,
-  fmtHashrate,
   formatAgo,
   formatFiat,
   formatFiatAmount,
@@ -33,7 +31,6 @@ import {
   THEME_ORDER,
   uptimeCell,
   WORKER_COLUMNS,
-  xvbDecisionRows,
 } from "./logic.mjs";
 import { MineCartTrain } from "./minecart.mjs";
 import { OsUpdateControl, OsVerdictBanner } from "./osupdate.mjs";
@@ -42,46 +39,12 @@ import { SecurityPanel } from "./securityview.mjs";
 import { MoreStats, nodeLocation, StatCard, TariStatus } from "./statcards.mjs";
 import { StackTopology } from "./topology.mjs";
 import { WorkerInspect } from "./workerview.mjs";
+import { XvbTierBlock } from "./xvbview.mjs";
 
 // Palette token -> text-colour class (defined in dashboard.css).
 const cVar = (v) => "c-" + v;
 
 // --- Small shared pieces -------------------------------------------------------------
-
-// Standardized Day/Month/Year estimate table — the one shape every earnings tab presents its
-// estimate in: coin column in accent, a ≈-fiat column appearing once the coin's price is known
-// (the same visibility gate the old per-cell fiat cards used). Replaces per-tab stat-grids whose
-// two-column flow paired unrelated cells (XMR/year beside ≈/day). The coin triplet shares one
-// precision (coinTriplet) so the rows align.
-const EstTable = ({ unit, day, month, year, price, currency, title }) => {
-  const coins = coinTriplet([day, month, year], unit);
-  const haveFiat = Number.isFinite(price) && price > 0;
-  const rows = [
-    ["Day", day, coins[0]],
-    ["Month", month, coins[1]],
-    ["Year", year, coins[2]],
-  ];
-  return html`
-    <div class="est-scroll">
-    <table class="est-table" title=${title || ""}>
-        <thead><tr>
-            <th></th>
-            <th scope="col">${unit}</th>
-            ${haveFiat ? html`<th scope="col">≈ ${currency || "USD"}</th>` : null}
-        </tr></thead>
-        <tbody>
-            ${rows.map(
-              ([label, raw, coin]) => html`
-            <tr>
-                <th scope="row">${label}</th>
-                <td class="c-accent">${coin}</td>
-                ${haveFiat ? html`<td>${formatFiatAmount(coinFiat(raw, price))}</td>` : null}
-            </tr>`,
-            )}
-        </tbody>
-    </table>
-    </div>`;
-};
 
 // Sign colour for a net figure — the one judgment colour in the calculator: green when the
 // operation profits, red when it loses. Everything that is a plain estimate stays accent/plain.
@@ -451,146 +414,6 @@ function NetworkCard({ state }) {
     <div class="card card-advanced" id="card-network">
         <h3>XMR Network</h3>
         <${MoreStats} prefKey="dashboardCardNetwork" headline=${headline} detail=${detail} count=${8} />
-    </div>`;
-}
-
-// XvB tier decision table (#872, study-final): the analytical tool a miner decides with. Every
-// donor tier on one row — draw odds (live winners feed), cost at YOUR hashrate, XvB's published
-// face value AND the study estimate (face x the measured on-chain delivery band) side by side,
-// and a coloured net verdict. This wallet's own measured wins supersede the study band when they
-// exist. Unsustainable tiers stay visible but grayed with the net withheld. No dropdown: the
-// comparison IS the decision, so all tiers show at once.
-function XvbDecisionTable({ calc, coeffDay, hr, energy }) {
-  const rows = xvbDecisionRows(calc, coeffDay, hr);
-  if (!rows.length) return null;
-  const measured = calc.realization_pct !== null && calc.realization_pct !== undefined;
-  const fmtRange = (r) =>
-    r === null ? "—" : r[0] === r[1] ? formatXmr(r[0]) : `${formatXmr(r[0])} … ${formatXmr(r[1])}`;
-  const estHeader = measured
-    ? `Yours (${calc.realization_pct}% × ${calc.realization_wins} wins)`
-    : "Study est. / yr";
-  // Fiat mirror (#520) for the best sustainable net only — one line, never a fiat number whose
-  // XMR figure is hidden.
-  const best = rows.filter((r) => r.sustainable && r.net).sort((a, b) => b.net[1] - a.net[1])[0];
-  return html`
-        <div class="xvb-comparison">
-            <label class="xvb-compare-label">Should I donate? — per-tier verdict (per year)</label>
-            <p class="text-muted text-xs" id="xvb-study-note">
-                XvB's figures are face value. A 25-round single-wallet on-chain audit (Jun–Aug
-                2026, all three sidechains) measured winners receiving 33% of face (95% CI
-                28–39%), with at most a small margin effect; a 14-winner public crawl
-                corroborates (no winner near face value). The
-                ${measured ? "Yours" : "Study"} column prices that in; the Net verdict uses it.
-            </p>
-            <div class="table-scroll">
-            <table class="est-table" id="xvb-decision-table">
-                <thead><tr>
-                    <th>Tier</th>
-                    <th scope="col" title="How often this tier's rounds pay out, and among how many qualifiers — from XvB's public winners feed.">Odds / 30d</th>
-                    <th scope="col" title="P2Pool earnings forgone by donating the tier threshold for a year, at your current rate.">Cost / yr</th>
-                    <th scope="col" title="XvB's own published expected reward — face value: prices every bonus hash at full block reward.">XvB says / yr</th>
-                    <th scope="col" title=${
-                      measured
-                        ? "XvB's figure scaled by what THIS wallet's raffle wins actually paid."
-                        : "XvB's figure scaled by the measured delivery band (33%, CI 28–39% — on-chain single-wallet audit, 25 rounds; crawl-corroborated)."
-                    }>${estHeader}</th>
-                    <th scope="col" title="Estimated reward minus the P2Pool earnings given up. Red: loses even at the optimistic end. Green: profits even at the pessimistic end.">Net / yr</th>
-                </tr></thead>
-                <tbody>
-                ${rows.map((r) => {
-                  const est = r.yours !== null ? [r.yours, r.yours] : r.study;
-                  return html`<tr class=${r.sustainable ? "" : "text-muted"}
-                      title=${r.sustainable ? "" : `Not sustainable at your hashrate — holding this tier needs about ${fmtHashrate(r.threshold)} donated continuously.`}>
-                    <th scope="row">${r.name}${r.sustainable ? "" : " ⚠"}</th>
-                    <td>${r.oddsPer30d ? `≈ ${Number(r.oddsPer30d.toPrecision(2))} wins · ${Number((r.players || 0).toPrecision(2))} players` : "—"}</td>
-                    <td>${r.cost !== null ? formatXmr(r.cost) : "—"}</td>
-                    <td class="c-purple">${r.xvbSays !== null ? formatXmr(r.xvbSays) : "—"}</td>
-                    <td>${fmtRange(est)}</td>
-                    <td class=${r.sustainable ? r.cls : ""}>${r.sustainable ? fmtRange(r.net) : "—"}</td>
-                  </tr>`;
-                })}
-                </tbody>
-            </table>
-            </div>
-            ${
-              energy && energy.xmr_price > 0 && best
-                ? html`<p class="text-muted text-xs mt-1" id="xvb-fiat-line">
-                    ${best.name}: net ≈ ${formatFiat(coinFiat(best.net[0], energy.xmr_price), energy.currency)}
-                    … ${formatFiat(coinFiat(best.net[1], energy.xmr_price), energy.currency)} per year at the current XMR price</p>`
-                : null
-            }
-            <p class="text-muted text-xs mt-2">
-                ${
-                  calc.estimates_available
-                    ? "XvB figures fetched over Tor from the operator's published estimates; odds from the public winners feed; the draw is random among qualifiers — donating above a threshold buys no extra odds."
-                    : calc.estimates_source === "published"
-                      ? `Reward figures are XvB's last published table (${calc.estimates_published_date}), not a live fetch — XvB is off, so nothing is fetched from xmrvsbeast.com. Win odds need the live winners feed and stay unavailable until XvB runs again.`
-                      : "Expected reward estimate unavailable — tier costs only."
-                }
-            </p>
-        </div>`;
-}
-
-// XvB tier / raffle block (#118), inside the earnings card and driven by the same what-if
-// hashrate: the highest XMRvsBeast tier that hashrate sustains (computeXvbTier — the server's
-// own auto rule), what holding it costs, and the current vs target tier for context. Labelled
-// raffle status, never a payout. The draw is random above the threshold — donating more within
-// a tier buys nothing — but the odds themselves ARE knowable (#872: qualifier counts from XvB's
-// winners file), so the comparison dropdown shows them. Rendered with XvB disabled too (#938) —
-// the block is the enable/don't-enable decision aid — but the live-credit cards (Current/Target
-// tier) stand down with the flag: there is no credited donation to report.
-// `coeffDay` (earnings.coeff_day) feeds the per-tier payout comparison dropdown below.
-function XvbTierBlock({ calc, hr, coeffDay, energy, est }) {
-  if (!calc) return null;
-  const t = computeXvbTier(hr, calc);
-  return html`
-    <div class="xvb-tier-block">
-        <h4>XvB Tier (raffle)</h4>
-        ${
-          calc.enabled
-            ? null
-            : html`<p class="text-muted text-xs" id="xvb-disabled-note">
-                XvB donation is off — this table prices what enabling it would earn and cost at
-                your hashrate. Nothing is fetched from xmrvsbeast.com while it is off${
-                  calc.estimates_source === "published"
-                    ? html`, so the reward columns use XvB's last published table (${calc.estimates_published_date}) instead of a live read`
-                    : calc.estimates_source === "live"
-                      ? ", so the reward columns run from the last live read"
-                      : ""
-                }. The odds column needs the live winners feed, so it stays empty until
-                XvB runs again.</p>`
-        }
-        <div class="stat-grid">
-            <${StatCard} label="Sustainable Tier" value=${t ? t.tier : "None"} cls="c-purple"
-                         title="The highest XvB donor tier this hashrate sustains while leaving P2Pool its share — the same auto rule the donation controller uses." />
-            <${StatCard} label="Hashrate Cost" value=${t ? fmtHashrate(t.cost) : "—"}
-                         title="Holding the tier means continuously donating about its threshold — hashrate that earns no P2Pool shares while donated." />
-            ${
-              calc.enabled
-                ? html`
-            <${StatCard} label="Current Tier" value=${calc.current_tier}
-                         title="The tier your credited XvB donation clears right now (the lower of XvB's 1h and 24h averages)." />
-            <${StatCard} label="Target Tier" value=${calc.target_tier}
-                         title=${"The tier the donation controller is configured to aim for" + (calc.sustainable ? "." : " — currently NOT sustainable at your hashrate.")} />`
-                : null
-            }
-        </div>
-        ${
-          // Current-tier expected reward (#712) in the same standardized Day/Month/Year table
-          // every other tab shows. Only when the server sent a fresh estimate — never fabricated.
-          // The server tempers the published figure by measured delivery (#902): this wallet's
-          // measured win payouts when enough wins exist, else the study band's midpoint — never
-          // face value. A fixed figure, NOT scaled by the what-if hashrate; the heading says so.
-          est && est.xvbDay !== null
-            ? html`
-        <h4 class="est-heading" title="XvB's published expected reward for the tier your fleet holds now, tempered by measured delivery: scaled to what this wallet's wins measurably paid when enough wins exist, else by the midpoint of the measured delivery band (28–39% of face value). XvB's raw figure shows only in the decision table's own column. A raffle expectation across all qualifiers, not scaled by the what-if hashrate above.">
-            Current Tier Expected Reward — tempered by measured delivery</h4>
-        <${EstTable} unit="XMR" day=${est.xvbDay} month=${est.xvbMonth} year=${est.xvbYear}
-                     price=${energy ? energy.xmr_price : 0} currency=${energy ? energy.currency : "USD"} />`
-            : null
-        }
-        <${XvbDecisionTable} calc=${calc} coeffDay=${coeffDay} hr=${hr} energy=${energy} />
-        <p class="text-muted text-xs mt-2">${calc.note}${calc.mode_note ? " " + calc.mode_note : ""}</p>
     </div>`;
 }
