+++ dashboard/tests/frontend/components.test.mjs
@@ -596,71 +596,6 @@ test('EarningsCard provenance line reflects the live price feed (#520)', () => {
     assert.match(html, /USD 0\.000400/); // tiny XTM price keeps its precision
 });
 
-test('XvB decision table: all tiers at once, study column, coloured net verdict (#872)', () => {
-    const base = clone();
-    base.earnings.available = true;
-    base.earnings.coeff_day = 1e-7;
-    base.earnings.p2pool_hr = 200000;
-    base.earnings.p2pool_hr_str = '200.00 kH/s';
-    base.xvb_calc = {
-        enabled: true, max_fraction: 0.85, estimates_available: true, estimates_stale: false,
-        current_tier: 'None', target_tier: 'Whale (100.00 kH/s+)', target_threshold: 100000,
-        sustainable: true, note: 'An XvB tier is raffle status, not an XMR payout.',
-        mode_note: null, realization_pct: null, realization_wins: null,
-        tiers: [
-            { name: 'Vip (10.00 kH/s+)', threshold: 10000, expected_reward_year: 0.81,
-              realized_reward_year: null, assumed_reward_year_range: [0.81 * 0.27, 0.81 * 0.38],
-              win_odds_day: 0.12, players_avg: 31.4 },
-            { name: 'Whale (100.00 kH/s+)', threshold: 100000, expected_reward_year: 6.17,
-              realized_reward_year: null, assumed_reward_year_range: [6.17 * 0.27, 6.17 * 0.38],
-              win_odds_day: 0.84, players_avg: 8.2 },
-            { name: 'Mega (1.00 MH/s+)', threshold: 1000000, expected_reward_year: 56.9,
-              realized_reward_year: null, assumed_reward_year_range: [56.9 * 0.27, 56.9 * 0.38],
-              win_odds_day: 9.4, players_avg: 1.0 },
-        ],
-    };
-    const up = renderApp({ state: base });
-    // One table, no dropdown, every tier a row with odds, both estimates and a verdict.
-    assert.match(up, /id="xvb-decision-table"/);
-    assert.doesNotMatch(up, /id="xvb-tier-select"/);
-    assert.match(up, /XvB says \/ yr/);
-    assert.match(up, /Study est\. \/ yr/);
-    assert.match(up, /winners receiving 33% of face/);
-    // Whale row: cost 3.65; study band 1.6659…2.3446; net band negative at both ends -> red.
-    assert.match(up, /1\.6659[0-9]* XMR … 2\.3446[0-9]* XMR/);
-    assert.match(up, /-1\.98[0-9]* XMR … -1\.30[0-9]* XMR/);
-    // Mega is unsustainable at 200k×0.85: flagged, net withheld.
-    assert.match(up, /Mega \(1\.00 MH\/s\+\) ⚠/);
-    // Draw odds render per row.
-    assert.match(up, /≈ 25 wins · 8\.2 players/);
-    // XvB's face value stays visible as XvB's own number.
-    assert.match(up, /6\.1700 XMR/);
-});
-
-test('XvB decision table: local measured wins supersede the study column (#872)', () => {
-    const base = clone();
-    base.earnings.available = true;
-    base.earnings.coeff_day = 1e-7;
-    base.earnings.p2pool_hr = 200000;
-    base.earnings.p2pool_hr_str = '200.00 kH/s';
-    base.xvb_calc = {
-        enabled: true, max_fraction: 0.85, estimates_available: true, estimates_stale: false,
-        current_tier: 'Whale (100.00 kH/s+)', target_tier: 'Whale (100.00 kH/s+)',
-        target_threshold: 100000, sustainable: true, note: 'raffle status', mode_note: null,
-        realization_pct: 32, realization_wins: 9,
-        tiers: [
-            { name: 'Whale (100.00 kH/s+)', threshold: 100000, expected_reward_year: 6.17,
-              realized_reward_year: 6.17 * 0.32, assumed_reward_year_range: null,
-              win_odds_day: 0.84, players_avg: 8.2 },
-        ],
-    };
-    const up = renderApp({ state: base });
-    assert.match(up, /Yours \(32% × 9 wins\)/);
-    assert.match(up, /1\.9744[0-9]* XMR/); // 6.17 × 0.32
-    assert.match(up, /-1\.6756[0-9]* XMR/); // net = 1.9744 − 3.65, single point
-});
-
-
 test('CadenceCard shows the — placeholders on a cold stack, real figures when available (#84)', () => {
     // The base fixture has no pool difficulty → cadence.available === false → server-sent dashes.
     const cold = renderApp();
@@ -680,79 +615,6 @@ test('CadenceCard shows the — placeholders on a cold stack, real figures when 
     assert.match(up, /1h 0m/);
     assert.match(up, /123%/);
     assert.match(up, /1,234,567/);
-});
-
-test('XvB decision table: verdict colours cover green and zero-spanning bands, stale degrades honestly (#872)', () => {
-    const base = clone();
-    base.earnings.available = true;
-    base.earnings.coeff_day = 1e-9; // tiny cost so a positive band is constructible
-    base.earnings.p2pool_hr = 200000;
-    base.earnings.p2pool_hr_str = '200.00 kH/s';
-    base.xvb_calc = {
-        enabled: true, max_fraction: 0.85, estimates_available: true, estimates_stale: false,
-        current_tier: 'None', target_tier: 'Vip (10.00 kH/s+)', target_threshold: 10000,
-        sustainable: true, note: 'raffle status', mode_note: null,
-        realization_pct: null, realization_wins: null,
-        tiers: [
-            // cost = 10000 × 1e-9 × 365 = 0.00365; band well above -> green at both ends
-            { name: 'Vip (10.00 kH/s+)', threshold: 10000, expected_reward_year: 0.81,
-              realized_reward_year: null, assumed_reward_year_range: [0.22, 0.31],
-              win_odds_day: 0.12, players_avg: 31.4 },
-            // band straddles cost -> neutral (no class)
-            { name: 'Whale (100.00 kH/s+)', threshold: 100000, expected_reward_year: 6.17,
-              realized_reward_year: null, assumed_reward_year_range: [0.03, 0.05],
-              win_odds_day: 0.84, players_avg: 8.2 },
-        ],
-    };
-    const up = renderApp({ state: base });
-    assert.match(up, /class="status-ok">0\.21635/); // green: even pessimistic end profits
-    assert.match(up, /class="">-0\.0065/); // zero-spanning: neutral cell class
-    // Stale estimates: costs still render, estimate/net columns dash, footer says costs only.
-    const stale = clone();
-    stale.earnings.available = true;
-    stale.earnings.coeff_day = 1e-7;
-    stale.earnings.p2pool_hr = 200000;
-    stale.xvb_calc = { ...base.xvb_calc, estimates_available: false, estimates_stale: true,
-        tiers: base.xvb_calc.tiers.map((t) => ({ ...t, expected_reward_year: null,
-            assumed_reward_year_range: null, realized_reward_year: null })) };
-    const sh = renderApp({ state: stale });
-    assert.match(sh, /tier costs only/);
-    assert.match(sh, /0\.365\d* XMR/); // vip cost at 1e-7
-});
-
-test('XvB decision table: the vendored-fallback reward columns render values and say so, not the bare "tier costs only" footer (#1214)', () => {
-    // #1214: with XvB disabled and never enabled (no live cache), the server fills the reward
-    // columns from its published-table fallback and labels it estimates_source: 'published'
-    // plus a date. The table must show real figures (not dashes) and the footer must say the
-    // numbers are the last-published table, not a live fetch — never the old bare "tier costs
-    // only" wording, which reads as "we have no idea" when the real answer is available.
-    const s = clone();
-    s.earnings.available = true;
-    s.earnings.coeff_day = 1e-9;
-    s.earnings.p2pool_hr = 200000;
-    // expected_reward_year/assumed_reward_year_range mirror what build_xvb_calc's corrected
-    // fallback (views.XVB_PUBLISHED_REWARD_FALLBACK["donor_vip"] = 0.81, the archived file's
-    // PER-PLAYER row, not the much larger pool-total row) would actually emit for a disabled box.
-    s.xvb_calc = {
-        enabled: false, max_fraction: 0.85, estimates_available: false, estimates_stale: false,
-        estimates_source: 'published', estimates_published_date: '2026-08-10',
-        current_tier: 'Disabled', target_tier: 'Disabled', target_threshold: 10000,
-        sustainable: true, note: 'raffle status', mode_note: null,
-        realization_pct: null, realization_wins: null,
-        tiers: [
-            { name: 'Vip (10.00 kH/s+)', threshold: 10000, expected_reward_year: 0.81,
-              realized_reward_year: null, assumed_reward_year_range: [0.81 * 0.28, 0.81 * 0.39],
-              win_odds_day: null, players_avg: null },
-        ],
-    };
-    const up = renderApp({ state: s });
-    // XvB says / yr and Study est. / yr are no longer dashed.
-    assert.match(up, /0\.810000 XMR/); // c-purple "XvB says" cell, face value straight from the fallback
-    assert.match(up, /0\.226800 XMR … 0\.315900 XMR/); // Study est. band = fallback face × the SAME prior
-    assert.doesNotMatch(up, /tier costs only/);
-    assert.match(up, /last published table \(2026-08-10\)/);
-    // The odds column has no such fallback — same root cause, still honestly empty (#1214).
-    assert.match(up, /<td>—<\/td>/);
 });
 
 test('XvBStats greys the credited figures and flags the footer when the fetch is stale (#311)', () => {
